+++ source/hougfx/include/hou/gfx/vertex_buffer.inl
@@ -6,7 +6,7 @@ namespace hou
 {
 
 template <typename T, bool DynamicStorage>
-vertex_buffer_t<T, DynamicStorage>::vertex_buffer_t(uint size)
+vertex_buffer_t<T, DynamicStorage>::vertex_buffer_t(size_t size)
   : vertex_buffer(size * sizeof(T), DynamicStorage)
 {}
 
@@ -21,7 +21,7 @@ vertex_buffer_t<T, DynamicStorage>::vertex_buffer_t(const span<const T>& data)
 
 
 template <typename T, bool DynamicStorage>
-uint vertex_buffer_t<T, DynamicStorage>::get_size() const
+size_t vertex_buffer_t<T, DynamicStorage>::get_size() const
 {
   HOU_DEV_ASSERT(get_byte_count() % sizeof(T) == 0u);
   return get_byte_count() / sizeof(T);
@@ -41,7 +41,7 @@ typename vertex_buffer_t<T, DynamicStorage>::data_type
 template <typename T, bool DynamicStorage>
 typename vertex_buffer_t<T, DynamicStorage>::data_type
   vertex_buffer_t<T, DynamicStorage>::get_sub_data(
-    uint offset, uint element_count) const
+    size_t offset, size_t element_count) const
 {
   HOU_DEV_ASSERT(get_byte_count() % sizeof(T) == 0u);
   HOU_PRECOND(offset + element_count <= get_size());
@@ -69,7 +69,7 @@ void vertex_buffer_t<T, DynamicStorage>::set_data(const span<const T>& data)
 template <typename T, bool DynamicStorage>
 template <bool DS, typename Enable>
 void vertex_buffer_t<T, DynamicStorage>::set_sub_data(
-  uint offset, const span<const T>& data)
+  size_t offset, const span<const T>& data)
 {
   HOU_DEV_ASSERT(get_byte_count() % sizeof(T) == 0u);
   HOU_PRECOND(offset + data.size() <= get_size());
