+++ source/housys/src/hou/sys/file.cpp
@@ -124,8 +124,8 @@ void file::putc(char c)
 
 size_t file::gets(std::string& str)
 {
-  const char* retval
-    = fgets(const_cast<char*>(str.data()), str.size(), m_handle);
+  const char* retval = fgets(
+    const_cast<char*>(str.data()), narrow_cast<int>(str.size()), m_handle);
   update_flags();
   if(retval == nullptr)
   {
